refactored claptrapped ui to its own file, ui.py for ease of reuse, and POC prep for embedding more AP stuff into native game UI
