+++ worlds/borderlands_tps/archi_defs.py
@@ -1120,7 +1120,7 @@ item_data_table = {
     "Trap: Slippery":                              BLTPSArchiData("", 0, item_kind=trap),
     "Trap: Item Explosion":                        BLTPSArchiData("", 0, item_kind=trap),
     "Trap: Rubber Ducky":                    BLTPSArchiData("", 0, item_kind=trap),
-    "Trap: Clap-in-a-Box":                 BLTPSArchiData("", 0, item_kind=trap),
+    "Trap: Clap-in-the-Box":                 BLTPSArchiData("", 0, item_kind=trap),
     "Trap: Fling":                                 BLTPSArchiData("", 0, item_kind=trap),
     "Trap: Leaky Wallet":                          BLTPSArchiData("", 0, item_kind=trap),
     # "Trap Spawn: Assassins":                       BLTPSArchiData("", 0, item_kind=trap),
